+++ requirements.txt
@@ -1,3 +1,3 @@
-Django>=2.2.0
+Django>=2.2
 social-auth-app-django>=4.0.0
 python-jose>=3.2.0
